requirements.txt
google-genai
python-dotenv
moviepy
Pillow